requirements.txt
fastapi
uvicorn[standard]
openai
pyyaml
python-dotenv
python-multipart
requests
pipecat-ai[openai,silero]
pdfplumber
arize-phoenix
openinference-instrumentation-openai
opentelemetry-sdk
opentelemetry-exporter-otlp
